refact(api): move GoalRequest/GoalResponse/HealthResponse from intelligent_agent.py to schemas_agent.py (#5977)
